refactor: organized all the assets into folders and removed unnecessary and duplicated assets.
