+++ client/src/components/lobby/PlayerCard.tsx
@@ -1,5 +1,5 @@
 import { PlayerAttributes } from "../../../../types/types.ts";
-import cross from "../../assets/semitransparent-cross.svg";
+import cross from "../../assets/misc/Cross.svg";
 
 interface PlayerCardProps {
   player: PlayerAttributes;
@@ -23,7 +23,7 @@ const PlayerCard = (props: PlayerCardProps) => {
           <img
             src={cross}
             alt={"Player Eliminated"}
-            className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 z-20 w-40 scale-50"
+            className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 z-20 w-40 scale-50 opacity-50"
           />
         ) : (
           ""
