+++ client/src/components/player-screen/choose-move/ChoosePlayerMove.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import RockOption from "../../../assets/ChooseMove/RockOption.svg";
-import PaperOption from "../../../assets/ChooseMove/PaperOption.svg";
-import ScissorsOption from "../../../assets/ChooseMove/ScissorOption.svg";
+import RockOption from "../../../assets/choose-move/RockOption.svg";
+import PaperOption from "../../../assets/choose-move/PaperOption.svg";
+import ScissorsOption from "../../../assets/choose-move/ScissorOption.svg";
 import { Action } from "../../../../../types/types.ts";
 import MoveSelection from "./MoveSelection.tsx";
 import { socket } from "../../../App.tsx";
