+++ client/src/components/player-screen/waiting-screens/LoadingEffect.tsx
@@ -1,4 +1,4 @@
-import loading from "../../../assets/Loading.svg";
+import loading from "../../../assets/misc/Loading.svg";
 
 type LoadingEffectProps = {
   isOpponent: boolean;
